[day9] Fix tiny bug
The Alg sometimes returned an answer that was
1 different than the actual answer. The reason was
the fact that the TAILs first position was not added
to the +t_hist+ registry. Only at the very end of the
sum by adding +1. This would sometimes count the origin
position twice if the TAIL actually visits the origin
more than once. The origin position is now added to the
registry, which fixes the issue.